Fix SSR: split cycle page into dynamic client-only import
Move all date-dependent logic into cycle-page-client.tsx loaded
with next/dynamic ssr:false to avoid hydration mismatches and
prerender crashes from new Date().
